+++ unit-tests/state/dns.test.js
@@ -308,6 +308,86 @@ describe("dns", () => {
       );
     });
     describe("dns.records.schema", () => {
+      it("should have a valid preference when type is not mx", () => {
+        assert.isFalse(
+          craig.dns.records.preference.invalid({ type: "AAAA" }),
+          "it should be valid"
+        );
+      });
+      it("should have an valid preference when type is mx and is not in range", () => {
+        assert.isTrue(
+          craig.dns.records.preference.invalid({ type: "MX", preference: -10 }),
+          "it should be invalid"
+        );
+      });
+      it("should have a valid port when type is not srv", () => {
+        assert.isFalse(
+          craig.dns.records.port.invalid({ type: "AAAA" }),
+          "it should be valid"
+        );
+      });
+      it("should have an valid port when type is srv and is not in range", () => {
+        assert.isTrue(
+          craig.dns.records.port.invalid({ type: "SRV", port: -10 }),
+          "it should be invalid"
+        );
+      });
+      it("should have a valid protocol when type is not srv", () => {
+        assert.isFalse(
+          craig.dns.records.protocol.invalid({ type: "AAAA" }),
+          "it should be valid"
+        );
+      });
+      it("should have an valid protocol when type is srv and no value", () => {
+        assert.isTrue(
+          craig.dns.records.protocol.invalid({ type: "SRV" }),
+          "it should be invalid"
+        );
+      });
+      it("should have a valid priority when type is not srv", () => {
+        assert.isFalse(
+          craig.dns.records.priority.invalid({ type: "AAAA" }),
+          "it should be valid"
+        );
+      });
+      it("should have an valid priority when type is srv and is not in range", () => {
+        assert.isTrue(
+          craig.dns.records.priority.invalid({ type: "SRV", priority: -12 }),
+          "it should be invalid"
+        );
+      });
+      it("should have a valid service when type is not srv", () => {
+        assert.isFalse(
+          craig.dns.records.service.invalid({ type: "AAAA" }),
+          "it should be valid"
+        );
+      });
+      it("should have invalid service", () => {
+        assert.isTrue(
+          craig.dns.records.service.invalid({ type: "SRV" }),
+          "it should be valid"
+        );
+        assert.isTrue(
+          craig.dns.records.service.invalid({ type: "SRV", service: "mmm" }),
+          "it should be valid"
+        );
+      });
+      it("should have a valid weight when type is not srv", () => {
+        assert.isFalse(
+          craig.dns.records.weight.invalid({ type: "AAAA" }),
+          "it should be valid"
+        );
+      });
+      it("should have invalid weight", () => {
+        assert.isTrue(
+          craig.dns.records.weight.invalid({ type: "SRV" }),
+          "it should be valid"
+        );
+        assert.isTrue(
+          craig.dns.records.weight.invalid({ type: "SRV", weight: "mmm" }),
+          "it should be valid"
+        );
+      });
       it("should return correct groups for zones", () => {
         assert.deepEqual(
           craig.dns.records.dns_zone.groups(
@@ -414,6 +494,12 @@ describe("dns", () => {
           resource_group: "service-rg",
           plan: "standard",
         });
+      });
+      it("should return invalid when invalid description", () => {
+        assert.isTrue(
+          craig.dns.zones.description.invalid({ description: "@@@" }),
+          "it should be invalid"
+        );
       });
       it("should return if string is not valid", () => {
         assert.isTrue(
@@ -625,6 +711,22 @@ describe("dns", () => {
         [],
         "it should add zone"
       );
+    });
+    describe("custom resolver schema", () => {
+      it("should be invalid when more than three subnets are selected", () => {
+        assert.isTrue(
+          craig.dns.custom_resolvers.subnets.invalid({ subnets: [1, 2, 3, 4] }),
+          "it should be invalid"
+        );
+      });
+      it("should be invalid when description is invalid", () => {
+        assert.isTrue(
+          craig.dns.custom_resolvers.description.invalid({
+            description: "[1, 2, 3, 4]",
+          }),
+          "it should be invalid"
+        );
+      });
     });
   });
 });